# Conflicts:
#	src/main/java/com/hunt2hand/controller/ProductoController.java
#	src/main/java/com/hunt2hand/repository/PerfilRepository.java
#	src/main/java/com/hunt2hand/repository/ProductoRepository.java
#	src/main/java/com/hunt2hand/repository/UsuarioRepository.java

+++ src/main/java/com/hunt2hand/repository/UsuarioRepository.java
@@ -7,6 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+
     Optional<Usuario> findTopByUsername(String username);
+
+
 }
